refactor(execution): split task driver
